requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
requests==2.31.0
plotly==5.18.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
google-generativeai==0.3.2
